Move some files around to simplify file structure
