+++ components/policies/PolicySearchBar.js
@@ -16,7 +16,7 @@
 
 import React from "react";
 import PropTypes from "prop-types";
-import SearchBar from "components/shared/search/SearchBar";
+import SearchBar from "components/shared/SearchBar";
 import { usePolicies } from "providers/policies";
 import { policyActions } from "reducers/policies";
 import { SEARCH_ALL } from "utils/constants";
