+++ test/components/shared/DetailsHeader.spec.js
